requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.4.6
python-dotenv==1.0.1
pydantic==2.7.1
python-multipart==0.0.9
bcrypt==4.1.3
PyJWT==2.8.0
